feat(app): change name to git_diff_parser
