+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in git-diff-parser.gemspec
+# Specify your gem's dependencies in git_diff_parser.gemspec
 gemspec
 gem 'byebug' if RUBY_VERSION >= '2.0.0'
 gem 'pry'
