--- lib/git/diff/parser/version.rb
@@ -1,7 +1,0 @@
-module Git
-  module Diff
-    module Parser
-      VERSION = '1.0.1-beta'
-    end
-  end
-end
