+++ lib/git_diff_parser/version.rb
@@ -1,0 +1,3 @@
+module GitDiffParser
+  VERSION = '1.0.1-beta'
+end
