+++ spec/git_diff_parser/patch_spec.rb
@@ -1,10 +1,10 @@
 require 'spec_helper'
 
-describe Git::Diff::Parser::Patch do
+describe GitDiffParser::Patch do
   describe '#changed_lines' do
     it 'returns lines that were modified' do
       patch_body = File.read('spec/support/fixtures/patch.diff')
-      patch = Git::Diff::Parser::Patch.new(patch_body)
+      patch = GitDiffParser::Patch.new(patch_body)
 
       expect(patch.changed_lines.size).to eq(3)
       expect(patch.changed_lines.map(&:number)).to eq [14, 22, 54]
@@ -13,7 +13,7 @@ describe Git::Diff::Parser::Patch do
 
     context 'when body is nil' do
       it 'returns no lines' do
-        patch = Git::Diff::Parser::Patch.new(nil)
+        patch = GitDiffParser::Patch.new(nil)
 
         expect(patch.changed_lines.size).to eq(0)
       end
